test/test-accesstoken-pkg/py/pyproject.toml
[project]
name = "test-accesstoken-pkg-py"
version = "1.0.0"
description = "Verify imbrace (PyPI) — auth via Access Token — exercises every ai_agent method"
requires-python = ">=3.9"
dependencies = [
  "imbrace==1.0.4",
  "python-dotenv>=1.0.0",
]
